requirements.txt
openai>=1.40.0
python-telegram-bot>=21.0
Pillow>=10.0.0
